requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
colorama==0.4.6
discord.py==2.3.1
Django==4.2.3
Flask==2.3.2
frozenlist==1.4.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
requests==2.31.0
soupsieve==2.4.1
South==1.0.2
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
Voice==0.1.0
Werkzeug==2.3.6
yarl==1.9.2
youtube-dl==2021.12.17
